Break shared_ptr cyclic dependency
Allows for Sequencer and deeper objects to be destroyed
when an Mpc is. This was not an issue in normal use cases of
the mpc library, but it was an issue when running unit tests,
because multiple tests run within the same process, so much of
the memory was never cleared until the mpc-tests process
exited. This ended up requiring absurd amounts of memory for
running all the tests in one go.