Fix stereo resampling
When loading a WAV of unsupported samplerate, it will be resampled.
Moreover, there is the native MPC2000XL functionality to resample.

In both cases, stereo sample data was treated as mono by
libsamplerate. Since the sampledata is not interleaved, there is
only the problem of not resetting the resampler state at the point
where the left channel data ends and the right channel data begins.

But of course this is bad and incorrect enough. However, we can't
just pass the number of channels to libsamplerate, because the
sampledata is interleaved. So we must process stereo sampledata
twice.