Improve plugin host synchronization
The main change here is the addition of the
mpc::sequencer::ExternalClock class. The idea
of this class is to make use of parameters
provided by most plugin hosts:

* ppqPosition -- The musical time in quarter
notes at frame 0 of a buffer

* ppqPositionOfLastBarStart -- Not entirely
sure what this can be used for, but I'm using
it to know whether the host's transport has
moved back in time due to looping.

Tests were added to keep number of produced
ticks by this clock in check. These tests are
based on data recorded from Ableton Live.

Other changes:

* FrameSeq/FrameSeqBase were merged into FrameSeq.

* mpc::Mpc::init() does not take the useless input
and output count parameters anymore.