Derive program name from data sometimes #249
See https://github.com/izzyreal/mpc/issues/249#issuecomment-2429637366

In summary, if the name encoded inside the file is a case-
insensitive equivalent of the file name, the name encoded inside
the file is used. This is nice, because, as opposed to the file name,
the encoded name can have mixed casing. The file name is all upper
case.