Fix linked channel display in mixer screens
We have a similar commit, but that was for the case when switching
tabs while in the MIXER screen already. This fix is for leaving
the MIXER screen in a channel-linked state, and going back again.
It was not visualizing the fact that channel-link is still active.
I checked against the real 2000XL, and indeed channel-link should
remain active in this case, so in this commit we make sure that
the mixer's appearance is congruent with link state when reopening
the MIXER screen.