MPC-259 Play full sequence in song mode
Songs were playing all but the last tick of each sequence. This
caused obvious problems with synchronization, apart from not
triggering any events on the last tick of a sequence (but it's
quite rare that users have an event there).

Also reduced cyclomatic complexity by writing an early return
statement.

Also added a comment about the return value of `processSongMode`.