Rename Clock classes
We rename ExternalClock to Clock, and OldClock to LegacyClock.

The original idea behind ExternalClock was that it would only
consume values produced by a plugin host. Since recently this
class also consumes values produced by the standalone host, so
rather than an external clock it is just a clock.

This prepares for moving the pulse generation code, which
currently resides in various places that consume Clock, into this
class.

This in turn prepares for retiring OldClock, which has been
renamed to the more accurate LegacyClock.

So the idea is that Clock will take care of all clock needs for
both standalone ("internal") and plugin ("external") use cases.