Fix recording with metronome only
There were a lot of issues with this kind of recording, by which
we refer to 2 types of recording where the metronome is playing,
even though the sequencer is not running. This can be done in the
Step Editor by pressing one or more pads, as well as in the Main
screen, by pressing one or more pads while holding REC.

Additionally some remaining issues are fixed where
Sequencer::move was invoked, assuming the parameter was still in
ticks (but it's in quarter notes now).