Fix song loop persistence
A long, long time ago, when Izmar was still a wee lad, with
programming and reasoning skills of equal propertions, that the
sin of all sins was committed. The property that dictates whether
a song should loop or not, was added both to the `SongScreen`,
yielding a global value that pertains to all songs, as well as to
the `Song`.

Long story short, it belongs in the `Song`. This commit removes
the global property and fixes the ALL file parser routines.

This commit also addresses the direct-to-disk-recording scenario
where a song is set to loop, but the direct-to-disk-recorder
temporarily disables looping in order to keep the recording not
so infinite. When the recording process has finished, it is set
back to whatever value the user had set.