Fix Voice sample rate after engine change
There was an issue with the incorrect sample rate being used
at the start of the voice, yielding a double invocation to
`initializeSamplerateDependents`. This in turn caused an
incorrect pitch if the sample was played back on any of the
voices that were already used to play back at the old
sample rate the first time these voices would be assigned for
playback.