Clean up MidiSw impl
For some reason there were 40 available footswitch
assignments and the user could scroll horizontally in the
MIDISW screen.

I just checked on the real 2000XL and it has only 4 available
assignments and there's no horizontal scrolling.

I probably added this a long time ago as an attempt to implement
full MIDI controllability, but I'm not sure. I couldn't find
any documentation or references regarding 40-instead-of-4.

Either way, the original doesn't have this, and recent reverse
engineering of the ALL file format confirms that there's only 4
assignments. See
https://github.com/izzyreal/mpc2000xl_kaitai/blob/b29f8333cf161342e8ae12b79d1dea8f894e3921/mpc2000xl.all.ksy#L329

So, in this commit, we get rid of the 40 and make it 4 and
remove horizontal scrolling.